# Rename - EmploymentAgency.Class => EmploymentAgency.Doman - Вынос классов в отдельную папку
