+++ EmploymentAgency/EmploymentAgency.Domain/Models/Applicant.cs
@@ -1,4 +1,4 @@
-﻿namespace EmploymentAgency.Class;
+﻿namespace EmploymentAgency.Domain.Models;
 
 /// <summary>
 ///  Класс для соискателя работы, наследуемый от User
