+++ EmploymentAgency/EmploymentAgency.Domain/Models/Employer.cs
@@ -1,4 +1,4 @@
-﻿namespace EmploymentAgency.Class;
+﻿namespace EmploymentAgency.Domain.Models;
 
 /// <summary>
 /// Класс для работодателя, наследуемый от User
