+++ EmploymentAgency/EmploymentAgency.Domain/Models/Resume.cs
@@ -1,4 +1,4 @@
-﻿namespace EmploymentAgency.Class;
+﻿namespace EmploymentAgency.Domain.Models;
 /// <summary>
 /// Класс резюме
 /// </summary>
