+++ EmploymentAgency/EmploymentAgency.Domain/Models/Vacancy.cs
@@ -1,4 +1,4 @@
-﻿namespace EmploymentAgency.Class;
+﻿namespace EmploymentAgency.Domain.Models;
 
 /// <summary>
 /// Класс Вакансия.
